refactor: move common logic for renaming nodes on the workflow package
